requirements_old.txt
Jinja2>=3.0.0,<4.0.0
jmespath>=1.0.0,<2.0.0
pymodbus>=3.0.0,<4.0.0
setuptools>=68.0.0,<69.0.0
xsdata>=22.0.0,<23.0.0
aiohttp>=3.0.0,<4.0.0
sgr-lib>=0.0.2
certifi>=2023.11.17
cachetools>=5.3.2
PyYAML>=6.0.1


